requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
motor
passlib[bcrypt]
python-jose[cryptography]
email-validator
chromadb[persistent]
transformers
torch
pydantic-settings
